Fixing iterative query. Now it seem to be fully working in async mode
The interesting observation.  There could be two approaches to NDNS iterative query
verification.  In the one approach, each intermediate results is
independently verified and cached.  In another approach, only the final
result is verified.  The outcome is kind of the same, the final result
will be either accepted or discarded, however, the first is more
"secure", as it can abort query process earlier.  Of course, the
drawback is that it can perform more verification steps (though
generally, number of steps should be the same).